+++ src/EventSubscriber/LogoutEventSubscriber.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace Umanit\SamlBundle\EventSubscriber;
 
+use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Security\Http\Event\LogoutEvent;
 use Umanit\SamlBundle\Enums\Mode;
+use Umanit\SamlBundle\Exception\ProviderDisabledException;
+use Umanit\SamlBundle\Exception\ProviderNotFoundException;
 use Umanit\SamlBundle\Security\Http\Authenticator\Token\SamlToken;
 use Umanit\SamlBundle\Service\ConfigurationServiceInterface;
 use Umanit\SamlBundle\Service\SloServiceInterface;
@@ -15,9 +18,9 @@ final class LogoutEventSubscriber implements EventSubscriberInterface
 {
     public function __construct(
         private readonly ConfigurationServiceInterface $configurationService,
-        private readonly SloServiceInterface $sloService
+        private readonly SloServiceInterface $sloService,
+        private readonly LoggerInterface $logger,
     ) {
-
     }
 
     public static function getSubscribedEvents(): array
@@ -37,15 +40,27 @@ final class LogoutEventSubscriber implements EventSubscriberInterface
 
         $provider = $token->getProviderKey();
         $user = $token->getUser();
-        $configuration = $this->configurationService->getByProvider($provider);
+
+        try {
+            $configuration = $this->configurationService->getByProvider($provider);
+            // @formatter:off
+        } catch (ProviderNotFoundException | ProviderDisabledException) {
+            // @formatter:on
+            return;
+        } catch (\Throwable $e) {
+            $this->logger->error('SAML Authentication getting logout failed', ['exception' => $e]);
+
+            return;
+        }
 
         if (true !== ($configuration['enable_slo'] ?? false)) {
             return;
         }
 
-        if ($configuration['type'] === Mode::SP_INITIATED) {
+        if (Mode::SP_INITIATED === $configuration['type']) {
             // SP initiated logout
             $response = $this->sloService->sendLogoutRequest($provider, $user);
+
             $event->setResponse($response);
             $event->stopPropagation();
 
@@ -53,7 +68,8 @@ final class LogoutEventSubscriber implements EventSubscriberInterface
         }
 
         // IDP initiated logout
-        $response = $this->sloService->sendLogoutResponse($provider, $user);
+        $response = $this->sloService->sendLogoutResponse($provider, $event->getRequest());
+
         $event->setResponse($response);
     }
 }