+++ CHANGELOG.md
@@ -7,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed
+### Added
 
 - Exception spécifique en cas de provider introuvable ou désactivé
 
 ### Changed
 
 - Un provider introuvable ou désactivé ne logue plus d’erreur
+
+### Fixed
+
+- Protection de la clef de cache du `ResponseValidator`
 
 ## [1.0.32] - 2026-04-23